src/fastapi_service/requirements.txt
fastapi
uvicorn[standard]
numpy
pillow
tritonclient[grpc]
faiss-cpu
pydantic
